Shift index page to component
