+++ components/IndexMain/styles.tsx
